Move ObsoleteLetStatementInspectionTests for QuickFixes into relevant files under RD.QuickFixes
