+++ RubberduckTests/QuickFixes/RemoveExplicitLetStatementQuickFixTests.cs
@@ -1,0 +1,49 @@
+﻿using System.Linq;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using RubberduckTests.Mocks;
+using System.Threading;
+using Rubberduck.Inspections.Concrete;
+using Rubberduck.Inspections.QuickFixes;
+using Rubberduck.VBEditor.SafeComWrappers.Abstract;
+using RubberduckTests.Inspections;
+
+namespace RubberduckTests.QuickFixes
+{
+    [TestClass]
+    public class RemoveExplicitLetStatementQuickFixTests
+    {
+
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void ObsoleteLetStatement_QuickFixWorks()
+        {
+            const string inputCode =
+@"Public Sub Foo()
+    Dim var1 As Integer
+    Dim var2 As Integer
+    
+    Let var2 = var1
+End Sub";
+
+            const string expectedCode =
+@"Public Sub Foo()
+    Dim var1 As Integer
+    Dim var2 As Integer
+    
+    var2 = var1
+End Sub";
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new ObsoleteLetStatementInspection(state);
+            var inspector = InspectionsHelper.GetInspector(inspection);
+            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
+
+            new RemoveExplicitLetStatementQuickFix(state).Fix(inspectionResults.First());
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+
+    }
+}
